:sparkles: enhancement: #1 toast feature add. listview selection single.

+++ BreakTimeApp/BreakTimeApp/ViewModels/Pages/NotifyViewModel.cs
@@ -4,6 +4,7 @@ using BreakTimeApp.Services;
 using BreakTimeApp.ViewModels.Windows;
 using BreakTimeApp.Views.Windows;
 using Microsoft.Extensions.Logging;
+using Microsoft.Toolkit.Uwp.Notifications;
 using System.Collections.ObjectModel;
 using Wpf.Ui.Controls;
 
@@ -32,14 +33,59 @@ namespace BreakTimeApp.ViewModels.Pages
             // Note: TimeStoreItemDataService(及びDbContext)はスコープが違うのでインジェクションができない
             _dataService = new TimeStoreItemDataService(new TimeStoreItemDbContext());
             _windowsProviderService = windowsProviderService;
+            _items = new ObservableCollection<TimeStoreItem>();
             _ = LoadItemsAsync();
+            ToastEventHandler();
+        }
+
+        private void ToastEventHandler()
+        {
+            // TODO: トースト処理を別クラスに記述
+            // トースト通知がアクティブ化されたときに呼び出されるイベントハンドラを設定
+            ToastNotificationManagerCompat.OnActivated += toastArgs =>
+            {
+                // トースト通知のアクティブ化の詳細を取得
+                ToastArguments args = ToastArguments.Parse(toastArgs.Argument);
+
+                Application.Current.Dispatcher.Invoke(async () => {
+                    // TimeStoreItemのguidからItemを取得
+                    string guid = args[Properties.ToastResources.guid];
+                    TimeStoreItem? item = Items.FirstOrDefault(findItem => findItem.ID.ToString() == guid);
+                    if (item == null) return Task.CompletedTask;
+
+                    // トースト通知がアクティブ化されたときの処理
+                    try
+                    {
+                        string action = args[Properties.ToastResources.action];
+                        if (action.Equals(Properties.ToastResources.accept))
+                        {
+                            await OnRemoveItem(item);
+                        }
+                        else if (action.Equals(Properties.ToastResources.snooze) ||
+                        action.Equals(Properties.ToastResources.dismiss))
+                        {
+                            item.Progress = 0;
+                            item.Span = TimeSpan.FromSeconds(item.MaxProgress);
+                            item.Timer.Tick += item.Timer_Tick;
+                            item.Timer.Tick += item.Toast_ElapsedEventHandler;
+                            item.IsRunning = action == Properties.ToastResources.dismiss;
+                            OnToggleItem(item);
+                        }
+                    }
+                    catch
+                    {
+                        await OnRemoveItem(item);
+                    }
+
+                    return Task.CompletedTask;
+                });
+            };
         }
 
         [LogAspect]
         private async Task LoadItemsAsync()
         {
             var itemsDb = await _dataService.GetAllTimeStoreItemsAsync();
-            Items = new ObservableCollection<TimeStoreItem>();
             foreach (var itemDb in itemsDb)
             {
                 TimeStoreItem item = TimeStoreItemToDbConverter.ConvertBack(itemDb);
@@ -57,21 +103,26 @@ namespace BreakTimeApp.ViewModels.Pages
 
         [LogAspect]
         [RelayCommand]
-        private async Task OnAddItem()
+        private void OnAddItem()
         {
             // 新しいアイテムを追加
-            TimeStoreItem newItem = new TimeStoreItem
-            {
-                ID = Guid.NewGuid(),
-                Name = $"{Properties.NotifyPageResources.item_title} {Items.Count + 1}",
-                Span = DEFAULT_TIMESPAN,
-                IsRunning = true,
-                Icon = SymbolRegular.TriangleRight20,
-                Progress = 0,
-                MaxProgress = DEFAULT_TIMESPAN.TotalSeconds
-            };
+            TimeStoreItem newItem = App.GetService<TimeStoreItem>();
+            newItem.ID = Guid.NewGuid();
+            newItem.Name = $"{Properties.NotifyPageResources.item_title} {Items.Count + 1}";
+            newItem.Span = DEFAULT_TIMESPAN;
+            newItem.IsRunning = true;
+            newItem.Icon = SymbolRegular.TriangleRight20;
+            newItem.Progress = 0;
+            newItem.MaxProgress = DEFAULT_TIMESPAN.TotalSeconds;
             // 破棄イベントの処理内容を登録
             newItem.Disposed += TimeStoreItem_Disposed;
+
+            AddItem(newItem);
+        }
+
+        [LogAspect]
+        private async Task AddItem(TimeStoreItem newItem)
+        {
             Items.Add(newItem);
 
             await _dataService.AddTimeStoreItemAsync(TimeStoreItemToDbConverter.Convert(newItem));
@@ -82,9 +133,13 @@ namespace BreakTimeApp.ViewModels.Pages
         [RelayCommand(CanExecute = nameof(canExecuteItem))]
         private async Task OnRemoveItem(TimeStoreItem item)
         {
-            Items.Remove(item);
-
+            // IDが同じものを調べる
+            var findItem = Items.FirstOrDefault(findItem => findItem.ID == item.ID);
+            if (findItem == null) return;
+            // 見つかった場合は削除する
+            Items.Remove(findItem);
             await _dataService.DeleteTimeStoreItemAsync(item.ID.ToString());
+
             _logger.LogInformation(AppLogEvents.UserAction, "item remove");
         }
 
@@ -92,7 +147,11 @@ namespace BreakTimeApp.ViewModels.Pages
         {
             if (sender is TimeStoreItem item)
             {
-                OnRemoveItem(item);
+                // UI表示からitemを削除する。(DBからは削除しない)
+                // ToastEventHandlerメソッドでitemの処理を決める。
+                // Items.Remove(item);
+                // DBとUI表示から削除
+                // OnRemoveItem(item);
             }
         }
